requirements.txt
bcrypt==4.1.3
blinker==1.8.2
cachelib==0.13.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Session==0.8.0
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==2.1.5
msgspec==0.18.6
numpy==2.0.0
opencv-python==4.10.0.84
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
openpyxl==3.1.5
gunicorn
gevent